backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI & ML
anthropic==0.8.1
openai==1.6.1
google-generativeai==0.3.2

# Google Cloud Vision OCR
google-cloud-vision==3.4.5
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-oauth2-tool==0.0.3

# PDF Processing
PyMuPDF==1.23.8
pdf2image==1.16.3
Pillow==10.1.0
contains-studio-agents==0.1.0
langchain==0.1.0
sentence-transformers==2.2.2
pgvector==0.2.3
numpy==1.24.3

# Image Processing & Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0

# Cloud Services
boto3==1.34.0  # AWS SQS
google-cloud-vision==3.5.0
sendgrid==6.11.0

# HTTP Client & Web Scraping
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
aiofiles==23.2.0

# Payment Processing
mercadopago==2.2.1

# Document Processing & Validation
PyPDF2==3.0.1
python-cpf==1.1.0
python-cnpj==1.0.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# WebSocket Support
websockets==12.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
httpx==0.25.2
aiosqlite==0.19.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0
bandit==1.7.5

# Monitoring & Logging
structlog==23.2.0
sentry-sdk==1.38.0

# File Processing
python-magic==0.4.27
Pillow==10.1.0

# Environment
python-dotenv==1.0.0
